include more info at the Main's homepage
Please reword as proper but this message is really missing and I think we should highlight it as much as possible. This is a first step.
I have met so many people at #GI2015 that knew Galaxy but did not understand any of these two concepts and it would benefit them very much.