WIP standardization/cleanup to more properly use the galaxy role:
- Use Gravity on Test
- Drop custom restarts
- Drop datamanager config

Unicornherder is not working properly as noted (even though it worked on Test before switching to Gravity). Gravity
would also maybe need to set at least $PATH and maybe $VIRTUAL_ENV, although the values as set in the systemd service
unit should suffice. Still, it would be useful to make it possible to set arbitrary env settings in Gravity.

For tusd under Gravity we need the ability to set $PATH since it's installed in /usr/local/bin.