Add x-clacks-overhead
This adds a non-standard header, that was inspired by the Discworld series. There a series of towers are used to constantly re-transmit the memory of a lost son, throughtout a network. Inspired by this many servers serve a similar header with memories of lost comrades (including sir terry pratchett, the author)